// src/app/api/model/discordVoiceChannelDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DiscordVoiceChannelDto { 
    guildId?: string;
    guildName?: string;
    channelId?: string;
    name?: string;
}